*** empty log message ***
git-svn-id: https://svn.cesecore.eu/svn/ejbca-restricted/trunk/ejbca@147 37aeb1de-b14b-0410-9a0e-82b28170d228

+++ src/java/se/anatom/ejbca/admin/ra.java
@@ -66,8 +66,13 @@ public class ra {
                     return;
                 }
                 String username = args[1];
-                System.out.println("Deleting user "+username);
-                admin.deleteUser(username);
+                System.out.print("Have you revoked the user [y/N]? ");
+                int inp = System.in.read();
+                if ( (inp == 121) || (inp==89) ) {
+                    System.out.println("Deleting user "+username);
+                    admin.deleteUser(username);
+                } else
+                    System.out.println("Delete aborted!");
             } else if (args[0].equals("setclearpwd"))
             {
                 if (args.length < 3) {
